requirements.txt
certifi==2020.4.5.1
cffi==1.14.0
cryptography==2.9.2
decorator==4.4.2
future==0.18.2
psycopg2==2.8.5
pycparser==2.20
python-dotenv==0.13.0
python-telegram-bot==12.6.1
six==1.14.0
tornado==6.0.4
